refactor: Graph 인터페이스 분리 및 빈 등록 사용
